fix(channels): drop unused draining-hold export and split its test
Knip flagged INGRESS_DRAINING_RELEASE_HOLD_MS as an unused production
export, and the extra case pushed ingress-drain.test.ts over max-lines.
Keep the constant module-private and move the hold regression into its
own file.
