+++ src/channels/message/ingress-drain.test.ts
@@ -1,7 +1,6 @@
 // Durable ingress drain contract tests for lifecycle reliability invariants.
 import { expectDefined } from "@openclaw/normalization-core";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import { GatewayDrainingError } from "../../process/gateway-work-admission.js";
 import { closeOpenClawStateDatabaseForTest } from "../../state/openclaw-state-db.js";
 import {
   createChannelIngressDrain,
@@ -167,46 +166,6 @@ describe("channel ingress drain", () => {
         expect(pending).toHaveLength(1);
         expect(pending[0]?.attempts).toBeGreaterThanOrEqual(1);
       });
-      drain.dispose();
-    });
-  });
-
-  it("holds a GatewayDrainingError claim before release so the next pump cannot re-claim instantly", async () => {
-    await withTempState(async (stateDir) => {
-      const queue = createTestIngressQueue(stateDir);
-      await queue.enqueue("evt-draining", { text: "x" }, { laneKey: "l1" });
-      const dispatches: string[] = [];
-      const drain = createChannelIngressDrain<Payload>({
-        queue,
-        dispatchClaimedEvent: async (event) => {
-          dispatches.push(event.id);
-          throw new GatewayDrainingError();
-        },
-      });
-
-      const idle = drain.waitForIdle();
-      await drain.drainOnce();
-      await vi.advanceTimersByTimeAsync(0);
-
-      expect(dispatches).toEqual(["evt-draining"]);
-      expect(await queue.listClaims()).toEqual([
-        expect.objectContaining({ id: "evt-draining", attempts: 0 }),
-      ]);
-      expect(await queue.listPending()).toEqual([]);
-      // Still claimed, so the next pump iteration cannot spin the same row.
-      expect(await drain.drainOnce()).toEqual({ started: 0 });
-
-      await vi.advanceTimersByTimeAsync(4_999);
-      expect(await queue.listClaims()).toHaveLength(1);
-      expect(await drain.drainOnce()).toEqual({ started: 0 });
-
-      await vi.advanceTimersByTimeAsync(1);
-      await idle;
-      expect(await queue.listClaims()).toEqual([]);
-      expect(await queue.listPending()).toEqual([
-        expect.objectContaining({ id: "evt-draining", attempts: 0 }),
-      ]);
-      expect(await queue.listFailed?.()).toEqual([]);
       drain.dispose();
     });
   });
